requirements.txt
aioredis==2.0.1
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
bidict==0.21.4
click==8.0.3
colorama==0.4.4
fastapi==0.73.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.3.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
pydantic==1.9.0
python-dotenv==0.19.2
python-engineio==4.3.1
python-socketio==5.5.1
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.0.1
uvicorn==0.17.4
watchgod==0.7
websockets==10.1
